@@ -1,6 +1,6 @@
 {
   "name": "react-dom",
-  "version": "17.0.2",
+  "version": "18.3.1",
   "description": "React package for working with the DOM.",
   "main": "index.js",
   "repository": {
@@ -18,17 +18,16 @@
   "homepage": "https://reactjs.org/",
   "dependencies": {
     "loose-envify": "^1.1.0",
-    "object-assign": "^4.1.1",
-    "scheduler": "^0.20.2"
+    "scheduler": "^0.23.2"
   },
   "peerDependencies": {
-    "react": "17.0.2"
+    "react": "^18.3.1"
   },
   "files": [
     "LICENSE",
     "README.md",
-    "build-info.json",
     "index.js",
+    "client.js",
     "profiling.js",
     "server.js",
     "server.browser.js",
@@ -37,9 +36,23 @@
     "cjs/",
     "umd/"
   ],
+  "exports": {
+    ".": "./index.js",
+    "./client": "./client.js",
+    "./server": {
+      "deno": "./server.browser.js",
+      "worker": "./server.browser.js",
+      "browser": "./server.browser.js",
+      "default": "./server.node.js"
+    },
+    "./server.browser": "./server.browser.js",
+    "./server.node": "./server.node.js",
+    "./profiling": "./profiling.js",
+    "./test-utils": "./test-utils.js",
+    "./package.json": "./package.json"
+  },
   "browser": {
-    "./server.js": "./server.browser.js",
-    "./unstable-fizz.js": "./unstable-fizz.browser.js"
+    "./server.js": "./server.browser.js"
   },
   "browserify": {
     "transform": [
